# storage/app/public/swagger/openapi.json
{
  "openapi" : "3.0.0",
  "info" : {
    "contact" : {
      "email" : "kosolapus@gmail.com",
      "name" : "Mike",
      "url" : "http://kosolapus.ddns.net/"
    },
    "description" : "API для будущей GroomCRM или назовите ее уже как-нибудь. На этой странице расписаны основные эндпоинты,\nпо которым можно получить данные из базы данных (или положить их туда, если будет такая возможность). Также здесь\nможно будет протестировать эти самые эндпоинты, посмотреть ответы и всякое такое.\n",
    "license" : {
      "name" : "MIT"
    },
    "title" : "Groomer Service",
    "version" : "1.0.2"
  },
  "servers" : [ {
    "url" : "http://kosolapus.ddns.net/api/v1"
  } ],
...
  "paths" : {
    "/oauth/client/login" : {
      "post" : {
        "description" : "Эндпоинт для авторизации клиента\n",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "example" : {
                "login" : "benjamin93@hotmail.com",
                "password" : "password"
              },
              "schema" : {
                "type" : "object"
              }
            }
          },
          "description" : "Аутентифицирует клиента - салон или частного грумера",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Bearer"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "Аутентифицирует клиента - салон или частного грумера\n",
        "tags" : [ "OAuth" ]
      }
    },
    "/oauth/master/login" : {
      "post" : {
        "description" : "Эндпоинт для авторизации Мастера\n",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "example" : {
                "login" : "libbie15@hotmail.com",
                "password" : "password"
              },
              "schema" : {
                "type" : "object"
              }
            }
          },
          "description" : "Аутентифицирует клиента - салон или частного грумера",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Bearer"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "Аутентифицирует клиента - салон или частного грумера\n",
        "tags" : [ "OAuth" ]
      }
    },
    "/client" : {
      "get" : {
        "description" : "Возвращает массив из Клиентов при их существовании или пустой\nмассив, если клиентов нет. Требует токен с уровнем прав как\nминимум администратора сервиса\n",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Client"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuthAdmin" : [ ]
        } ],
        "summary" : "Возвращает список клиентов массивом\n",
        "tags" : [ "Client" ]
      },
      "post" : {
        "description" : "Создает нового Клиента. Для работы с этим эндпоинтом\nтребуется доступ как минимум администратора сервиса\n",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Client"
              }
            }
          },
          "description" : "Store *Client* entity",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuthAdmin" : [ ]
        } ],
        "summary" : "Создает нового клиента\n",
        "tags" : [ "Client" ]
      }
    },
    "/client/{client}" : {
      "delete" : {
        "description" : "",
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuthAdmin" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Client" ]
      },
      "get" : {
        "description" : "",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Client"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Client" ]
      },
      "parameters" : [ {
        "description" : "Id requested Client",
        "example" : 1,
        "in" : "path",
        "name" : "client",
        "required" : true,
        "schema" : {
          "minimum" : 1,
          "type" : "integer"
        }
      } ],
      "patch" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Client"
              }
            }
          },
          "description" : "Optional description in *Markdown*",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuthAdmin" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Client" ]
      }
    },
    "/master" : {
      "get" : {
        "description" : "Store *Master* entity",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Id клиента, которому принадлежит мастер",
          "in" : "query",
          "name" : "client_id",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Master" ]
      },
      "post" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "allOf" : [ {
                  "properties" : {
                    "image_upload" : {
                      "format" : "binary",
                      "nullable" : false,
                      "type" : "string"
                    },
                    "name" : {
                      "nullable" : false,
                      "type" : "string"
                    },
                    "description" : {
                      "format" : "text",
                      "nullable" : true,
                      "type" : "string"
                    }
                  },
                  "required" : [ "description", "image_upload", "name" ],
                  "type" : "object"
                } ]
              }
            }
          },
          "description" : "Store *Master* entity",
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Master"
                }
              }
            },
            "description" : "Успешно создано"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_422"
                }
              }
            },
            "description" : "Невозможно обработать запрос, произошла ошибка в процессе"
          },
          "401" : {
            "description" : "Не передан токен"
          },
          "403" : {
            "description" : "Нет прав для выполнения данного действия"
          },
          "5XX" : {
            "description" : "Ошибка на сервере. Не должно быть таких ответов, это плохой признак"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Master" ]
      }
    },
    "/master/{master}" : {
      "delete" : {
        "description" : "",
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Master" ]
      },
      "get" : {
        "description" : "",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Master"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Master" ]
      },
      "parameters" : [ {
        "description" : "Id requested Master",
        "example" : 1,
        "in" : "path",
        "name" : "master",
        "required" : true,
        "schema" : {
          "minimum" : 1,
          "type" : "integer"
        }
      } ],
      "patch" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "$ref" : "#/components/schemas/Master"
              }
            }
          },
          "description" : "Optional description in *Markdown*",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Master" ]
      }
    },
    "/order" : {
      "get" : {
        "description" : "Store *Order* entity",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Order" ]
      },
      "post" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            }
          },
          "description" : "Store *Order* entity",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Order" ]
      }
    },
    "/order/{order}" : {
      "delete" : {
        "description" : "",
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Order" ]
      },
      "get" : {
        "description" : "",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Order"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Order" ]
      },
      "parameters" : [ {
        "description" : "Id requested Order",
        "example" : 1,
        "in" : "path",
        "name" : "order",
        "required" : true,
        "schema" : {
          "minimum" : 1,
          "type" : "integer"
        }
      } ],
      "patch" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Order"
              }
            }
          },
          "description" : "Optional description in *Markdown*",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Order" ]
      }
    },
    "/promotion" : {
      "get" : {
        "description" : "Store *Promotion* entity",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Id клиента, которому принадлежит мастер",
          "in" : "query",
          "name" : "client_id",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Promotion" ]
      },
      "post" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Promotion"
              }
            }
          },
          "description" : "Store *Promotion* entity",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Promotion" ]
      }
    },
    "/promotion/{promotion}" : {
      "delete" : {
        "description" : "",
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Promotion" ]
      },
      "get" : {
        "description" : "",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Promotion"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Promotion" ]
      },
      "parameters" : [ {
        "description" : "Id requested Promotion",
        "example" : 1,
        "in" : "path",
        "name" : "promotion",
        "required" : true,
        "schema" : {
          "minimum" : 1,
          "type" : "integer"
        }
      } ],
      "patch" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Promotion"
              }
            }
          },
          "description" : "Optional description in *Markdown*",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Promotion" ]
      }
    },
    "/push" : {
      "get" : {
        "description" : "Store *Push* entity",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Push" ]
      },
      "post" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Push"
              }
            }
          },
          "description" : "Store *Push* entity",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Push" ]
      }
    },
    "/push/{push}" : {
      "delete" : {
        "description" : "",
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Push" ]
      },
      "get" : {
        "description" : "",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Push"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Push" ]
      },
      "parameters" : [ {
        "description" : "Id requested Push",
        "example" : 1,
        "in" : "path",
        "name" : "push",
        "required" : true,
        "schema" : {
          "minimum" : 1,
          "type" : "integer"
        }
      } ],
      "patch" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Push"
              }
            }
          },
          "description" : "Optional description in *Markdown*",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Push" ]
      }
    },
    "/service" : {
      "get" : {
        "description" : "Store *Service* entity",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Id клиента, которому принадлежит мастер",
          "in" : "query",
          "name" : "client_id",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Service" ]
      },
      "post" : {
        "description" : "",
        "requestBody" : {
          "$ref" : "#/components/requestBodies/inline_object",
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "name" : {
                    "description" : "Название Услуги",
                    "type" : "string"
                  },
                  "image_upload" : {
                    "description" : "Загружаемое изображение услуги",
                    "format" : "binary",
                    "type" : "string",
                    "writeOnly" : true
                  },
                  "service_type" : {
                    "description" : "Тип услуги - для кошек, для собак, для других",
                    "items" : {
                      "enum" : [ 0, 1, 2 ],
                      "format" : "tinyint",
                      "type" : "integer"
                    },
                    "type" : "array"
                  }
                },
                "required" : [ "image_upload", "name", "service_type" ],
                "type" : "object"
              }
            }
          },
          "description" : "Store *Service* entity",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Service" ]
      }
    },
    "/service/{service}" : {
      "delete" : {
        "description" : "",
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Service" ]
      },
      "get" : {
        "description" : "",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Service"
                }
              }
            },
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "Service" ]
      },
      "parameters" : [ {
        "description" : "Id requested Service",
        "example" : 1,
        "in" : "path",
        "name" : "service",
        "required" : true,
        "schema" : {
          "minimum" : 1,
          "type" : "integer"
        }
      } ],
      "patch" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Service"
              }
            }
          },
          "description" : "Optional description in *Markdown*",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "Service" ]
      }
    },
    "/working-diapason" : {
      "get" : {
        "description" : "Store *WorkingDiapason* entity",
        "parameters" : [ {
          "description" : "Как много элементов должно возвращаться за один запрос",
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "type" : "integer"
          }
        }, {
          "description" : "Смещение от первого",
          "in" : "query",
          "name" : "offset",
          "required" : false,
          "schema" : {
            "default" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "Id мастера",
          "in" : "query",
          "name" : "master_id",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "summary" : "",
        "tags" : [ "WorkingDiapason" ]
      },
      "post" : {
        "description" : "",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkingDiapason"
              }
            }
          },
          "description" : "Store *WorkingDiapason* entity",
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Ok"
          }
        },
        "security" : [ {
          "bearerAuthMaster" : [ ]
        } ],
        "summary" : "",
        "tags" : [ "WorkingDiapason" ]
      }
    },
...
  },
  "components" : {
...
    "schemas" : {
...
      "Master" : {
        "description" : "Мастер - это непосредственный исполнитель услуги.\nРазные мастера могут предоставлять одну или несколько услуг из общего списка Клиента\nУ мастера есть расписание.\n",
        "properties" : {
          "id" : {
            "format" : "int64",
            "readOnly" : true,
            "type" : "integer"
          },
          "image" : {
            "format" : "url",
            "readOnly" : true,
            "type" : "string"
          },
          "image_upload" : {
            "format" : "binary",
            "type" : "string",
            "writeOnly" : true
          },
          "name" : {
            "nullable" : false,
            "type" : "string"
          },
          "description" : {
            "nullable" : true,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "Push" : {
        "description" : "TODO: Осмысленное описание объекта Пушей\n",
        "properties" : {
          "id" : {
            "format" : "int64",
            "readOnly" : true,
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Bearer" : {
        "description" : "Авторизационные токены.\n",
        "properties" : {
          "token_type" : {
            "description" : "Уникальный идентификатор Клиента",
            "readOnly" : true,
            "type" : "string"
          },
          "access_token" : {
            "description" : "Уникальный идентификатор Клиента",
            "readOnly" : true,
            "type" : "string"
          },
          "refresh_token" : {
            "description" : "Уникальный идентификатор Клиента",
            "readOnly" : true,
            "type" : "string"
          },
          "expires_in" : {
            "description" : "Уникальный идентификатор Клиента",
            "format" : "int64",
            "readOnly" : true,
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "inline_response_422" : {
        "description" : "Сообщение об ошибке",
        "properties" : {
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "bearerAuthMaster" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "bearerAuthAdmin" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}